+++ lib/executable.dart
@@ -1,0 +1,2 @@
+export 'src/executable/builder_config.dart';
+export 'src/executable/executable.dart';